Change SSM Parameter Store path to /iam/cis-publishers/
Previously the SSM Parameter Store parameters were stored in the root of /iam/ and not scoped to this product. This moves them into a scoped path. Once in a scoped path it will allow granular IAM permissions to Parameter Store parameters on a per product basis

Once this is merged and deployed, cleanup by
- [ ] Deleting the no longer used SSM Parameter Store parameters `/iam/development/ldap_signing_key`, `/iam/production/ldap_signing_key` and `/iam/testing/ldap_signing_key`