Remove the unused profile creation code into a dedicated file to prevent confusion
People (including me) repeatedly find the profile creation code and don't notice
in the page of code above the "return false" which causes the code to never be traversed.
This leads to confusion as readers think that the LDAP publisher creates ne profiles,
which it doesn't.

This moves the code into a names function that makes it clear it isn't used and adds
comments to point out that the code isn't used.